requirements.txt
# requirements.txt
# 项目依赖包列表

Flask>=2.0.0 # Flask Web框架
Flask-SQLAlchemy>=2.5.0 # Flask扩展，用于SQLAlchemy ORM
Flask-Migrate>=3.0.0 # Flask扩展，用于SQLAlchemy数据库迁移 (Alembic)
psycopg2-binary # PostgreSQL数据库驱动 (如果使用PostgreSQL)
# pyodbc # SQL Server数据库驱动 (如果使用SQL Server, 请取消注释并注释掉psycopg2-binary)
python-dotenv>=0.19.0 # 用于从.env文件加载环境变量

# Testing
pytest>=7.0.0 # Python测试框架
pytest-flask>=1.2.0 # Pytest插件，用于测试Flask应用
# Faker # (可选) 用于生成测试数据
# pytest-cov # (可选) 用于测试覆盖率报告
